+++ benchmark/graph_color/benchmark.jl
@@ -2,7 +2,7 @@ function solve_us_graph_coloring(; num_colors = 8, equality = false)
     m = Model(optimizer_with_attributes(
         CS.Optimizer,
         "logging" => [],
-        "branch_strategy" => :OLD,
+        "branch_strategy" => :IMPS,
         "seed" => 1,
     ))
     @variable(m, 1 <= max_color <= num_colors, Int)